requirements.txt
allure-pytest
selenium
chromedriver